p29.25 refactor: thread ids.FundID everywhere
Typed-ID capstone: FundID is now a distinct type end-to-end. Purely types, no
behavior change; report goldens stay byte-identical.

Threaded:
- sqlc: Fund.ID, FundSubsidiary.FundID, {Funds,FundSubsidiaries}Version.EntityID;
  every fund-id query param/return (GetFund, FundSubsidiaries, InsertFund return,
  Insert/Delete/HasFundSubsidiary(Map)Params.FundID, InsertFund(Subsidiary)Version
  entity ids, UpdateFundParams.ID); the report-row fund ids (FundBalancesAsOfRow,
  CurrentCashFundBalancesAsOfRow, BudgetKeyActivityRow .FundID) that COALESCE to 0
  = unrestricted.
- store: CreateFund/GetFund/UpdateFund/Close/Reopen/setFundActive/FundSubsidiary-
  IDs/FundLedger fund-id params+returns; add/remove/insertFundVersion + fundSub-
  SetTx + fundSplitInSubsidiary; domain FundID on FundCurrencyAmount, BudgetKeyCell,
  DescriptionPrefill; nullable *ids.FundID on SplitInput/BudgetSplitInput/Expense-
  ReportLineInput, RegisterFilters/Row, DrillFilter/Row, Budget/DrillRow, recon
  workspace/statement splits -- ids.Null on write, ids.Ptr[ids.FundID] on read.
- reports: reports.FundID alias already ids.FundID; typed the fund-keyed maps
  (map[FundID]string), fundCell/statementFundCell/fundLabelCell/fundIDForDrill/
  restrictedFundIDs/drillFunds/nonExpense+revenueWith+releasedDrill, Drill.FundID(s)
  + Encode/Decode (int64 at the URL string boundary).
- web: parse to typed at parseID seams; render-only model fields stay int64
  (int64(f.ID) at assignment); fund-keyed maps typed (fundNameMap, histResolver.
  funds, byFund); derefID made generic ~int64.
- CLI/synth/fixture: ledgerimport BuildResult.FundIDs/CampusFundID + residual/
  fundOf maps; synth IDs.{BecaAgua,BuildingFund,Campaign} + sp.fund/budget fund
  (alias-imported entids to dodge the `ids` IDs-param shadow); fixture expected
  fund fields.
- tests: fund helper params/env fields typed; generic helpers (key2/wantCell/nn/ni
  made ~int64; AssertVersioned(Fund)/itoa/fundStatementURL stay int64 -- callers
  pass int64(fundID)); URL/scs unchanged.

Boundary discipline (Rule 15): split-level nullable fund FKs stay sql.NullInt64 at
the scan boundary (Split(s)Version, ExpenseReportLine, budget_splits, history
DiffValue.ID heterogeneous holder), typed only at the store seam.

+++ internal/reports/capital_campaign.go
@@ -108,7 +108,7 @@ func runCapitalCampaign(ctx context.Context, tk *Toolkit, p Params) (Table, erro
 	// The fund's whole ledger to the report date: used ONLY for the earliest date (the
 	// span start) and the per-account capital-detail balances. The classification and the
 	// line-item figures come from FundStatement below, not from re-scanning these rows.
-	rows, err := tk.store.FundLedger(ctx, int64(p.Fund), p.To)
+	rows, err := tk.store.FundLedger(ctx, p.Fund, p.To)
 	if err != nil {
 		return Table{}, err
 	}
